Got ngspice working with pyspice
Okay, we now have a rough feasibility estimate. We can get about
100,000 samples per second for a trivially simple circuit using
pyspice. That said, I'm willing to believe that this would faster
if the thing feeding the samples wasn't in python.

Still, Human hearing caps out at around 20 KHz, and 44.1 KHz is the
standard sampling rate for audio files, so we should be able to
pull this off even in python.